% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdetidy-package.R
\docType{package}
\name{cdetidy-package}
\alias{cdetidy}
\alias{cdetidy-package}
\title{cdetidy: Tools for Working with CDE Datasets}
\description{
Utilities for classification, labeling, schema validation, and export of
public California Department of Education datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jaubele.github.io/cdetidy}
}

}
\author{
\strong{Maintainer}: Joseph Aubele \email{jaubele@ocde.us}

Authors:
\itemize{
  \item Joseph Aubele \email{jaubele@ocde.us}
  \item Caleb Mitchell
  \item Aubrey Roy
}

}
\keyword{internal}
